branch_exports/cursor/bc-1114e741-a737-4c43-b6bd-74bdd29ce1f8-e61a/requirements.txt
jsonschema==4.22.0
PyYAML==6.0.2